Add files via upload
The assignment was divided into two main parts:

Redis Implementation: We used Redis to efficiently manage and query the dataset related to seller actions. By implementing code in R, we optimized data retrieval processes, enabling us to analyze user engagement with email campaigns and determine the effectiveness of these campaigns on seller activity.

MongoDB Implementation: We used MongoDB to manage the motorcycle advertisements dataset. Python, along with the pymongo library, was used to clean, transform, and insert data into MongoDB. We performed complex queries to determine key metrics such as the average price of motorcycles, brand-specific insights, and feature-based analyses like the prevalence of ABS in motorcycles.

Through this assignment, we gained hands-on experience in handling big data systems and architectures, learned to integrate and analyze large-scale datasets using different database technologies, and developed skills in both R and Python for data manipulation and analysis.